pagemap: use unsigned long for page counts
Variables storing page counts were previously `unsigned int`, limiting
them to a maximum of 2^32 pages. With a 4k page size, this corresponds
to a 16TB memory mapping, which is insufficient for larger mappings.

This commit changes the type for these variables to `unsigned long` to
support larger memory mappings.

+++ criu/include/parasite.h
@@ -63,7 +63,7 @@ struct parasite_dump_pages_args {
 	unsigned int add_prot;
 	unsigned int off;
 	unsigned int nr_segs;
-	unsigned int nr_pages;
+	unsigned long nr_pages;
 };
 
 static inline struct parasite_vma_entry *pargs_vmas(struct parasite_dump_pages_args *a)
